pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "faiss-cpu>=1.10.0",
    "openai>=1.70.0",
    "pymupdf>=1.25.5",
    "streamlit>=1.44.1",
]
